+++ packages/pglite/src/pglite.ts
@@ -75,14 +75,12 @@ export class PGlite
   #notifyListeners = new Map<string, Set<(payload: string) => void>>()
   #globalNotifyListeners = new Set<(channel: string, payload: string) => void>()
 
-  static readonly RECV_BUF_SIZE: number = 16 * 1024 * 1024 // 16MB default
-
   // receive data from wasm
   #pglite_write: number = -1
-  // buffer that holds data received from wasm
-  #inputData = new Uint8Array(PGlite.RECV_BUF_SIZE)
-  // write index in the buffer
-  #writeOffset: number = 0
+
+  #currentResults: BackendMessage[] = []
+  #currentThrowOnError: boolean = false
+  #currentOnNotice: ((notice: NoticeMessage) => void) | undefined
 
   // send data to wasm
   #pglite_read: number = -1
@@ -90,6 +88,16 @@ export class PGlite
   #outputData: any = []
   // read index in the buffer
   #readOffset: number = 0
+  #currentDatabaseError: DatabaseError | null = null
+
+  #keepRawResponse: boolean = true
+  // these are needed for point 2 above
+  static readonly DEFAULT_RECV_BUF_SIZE: number = 1 * 1024 * 1024 // 1MB default
+  static readonly MAX_BUFFER_SIZE: number = Math.pow(2, 30)
+  // buffer that holds data received from wasm
+  #inputData = new Uint8Array(0)
+  // write index in the buffer
+  #writeOffset: number = 0
 
   /**
    * Create a new PGlite instance
@@ -379,24 +387,30 @@ export class PGlite
     this.mod = await PostgresModFactory(emscriptenOpts)
 
     // set the write callback
-    this.#pglite_write = (this.mod as any).addFunction(
-      (ptr: any, length: number) => {
-        let bytes
-        try {
-          bytes = this.mod!.HEAPU8.subarray(ptr, ptr + length)
-        } catch (e: any) {
-          console.error('error', e)
-          throw e
-        }
+    this.#pglite_write = this.mod.addFunction((ptr: any, length: number) => {
+      let bytes
+      try {
+        bytes = this.mod!.HEAPU8.subarray(ptr, ptr + length)
+      } catch (e: any) {
+        console.error('error', e)
+        throw e
+      }
+      this.#protocolParser.parse(bytes, (msg) => {
+        this.#parse(msg)
+      })
+      if (this.#keepRawResponse) {
         const copied = bytes.slice()
 
-        const requiredSize = this.#writeOffset + copied.length
+        let requiredSize = this.#writeOffset + copied.length
 
         if (requiredSize > this.#inputData.length) {
           const newSize =
             this.#inputData.length +
             (this.#inputData.length >> 1) +
             requiredSize
+          if (requiredSize > PGlite.MAX_BUFFER_SIZE) {
+            requiredSize = PGlite.MAX_BUFFER_SIZE
+          }
           const newBuffer = new Uint8Array(newSize)
           newBuffer.set(this.#inputData.subarray(0, this.#writeOffset))
           this.#inputData = newBuffer
@@ -406,34 +420,31 @@ export class PGlite
         this.#writeOffset += copied.length
 
         return this.#inputData.length
-      },
-      'iii',
-    )
+      }
+      return length
+    }, 'iii')
 
     // set the read callback
-    this.#pglite_read = (this.mod as any).addFunction(
-      (ptr: any, max_length: number) => {
-        // copy current data to wasm buffer
-        let length = this.#outputData.length - this.#readOffset
-        if (length > max_length) {
-          length = max_length
-        }
-        try {
-          this.mod!.HEAP8.set(
-            (this.#outputData as Uint8Array).subarray(
-              this.#readOffset,
-              this.#readOffset + length,
-            ),
-            ptr,
-          )
-          this.#readOffset += length
-        } catch (e) {
-          console.log(e)
-        }
-        return length
-      },
-      'iii',
-    )
+    this.#pglite_read = this.mod.addFunction((ptr: any, max_length: number) => {
+      // copy current data to wasm buffer
+      let length = this.#outputData.length - this.#readOffset
+      if (length > max_length) {
+        length = max_length
+      }
+      try {
+        this.mod!.HEAP8.set(
+          (this.#outputData as Uint8Array).subarray(
+            this.#readOffset,
+            this.#readOffset + length,
+          ),
+          ptr,
+        )
+        this.#readOffset += length
+      } catch (e) {
+        console.log(e)
+      }
+      return length
+    }, 'iii')
 
     this.mod._set_read_write_cbs(this.#pglite_read, this.#pglite_write)
 
@@ -565,6 +576,8 @@ export class PGlite
     try {
       await this.execProtocol(serialize.end())
       this.mod!._pgl_shutdown()
+      this.mod!.removeFunction(this.#pglite_read)
+      this.mod!.removeFunction(this.#pglite_write)
     } catch (e) {
       const err = e as { name: string; status: number }
       if (err.name === 'ExitStatus' && err.status === 0) {
@@ -643,13 +656,18 @@ export class PGlite
    * @returns The direct message data response produced by Postgres
    */
   execProtocolRawSync(message: Uint8Array) {
-    // let data
     const mod = this.mod!
-    // >0 set buffer content type to wire protocol
-    mod._use_wire(1)
 
-    if (this.#inputData.buffer.byteLength > PGlite.RECV_BUF_SIZE) {
-      this.#inputData = new Uint8Array(PGlite.RECV_BUF_SIZE)
+    this.#readOffset = 0
+    this.#writeOffset = 0
+    this.#outputData = message
+
+    if (
+      this.#keepRawResponse &&
+      this.#inputData.length !== PGlite.DEFAULT_RECV_BUF_SIZE
+    ) {
+      // the previous call might have increased the size of the buffer so reset it to its default
+      this.#inputData = new Uint8Array(PGlite.DEFAULT_RECV_BUF_SIZE)
     }
     this.#readOffset = 0
     this.#outputData = message
@@ -661,7 +679,8 @@ export class PGlite
 
     this.#outputData = []
 
-    if (this.#writeOffset) return this.#inputData.subarray(0, this.#writeOffset)
+    if (this.#keepRawResponse && this.#writeOffset)
+      return this.#inputData.subarray(0, this.#writeOffset)
     return new Uint8Array(0)
   }
 
@@ -700,14 +719,72 @@ export class PGlite
       onNotice,
     }: ExecProtocolOptions = {},
   ): Promise<ExecProtocolResult> {
-    const data = await this.execProtocolRaw(message, { syncToFs })
-    const results: BackendMessage[] = []
+    this.#currentThrowOnError = throwOnError
+    this.#currentOnNotice = onNotice
+    this.#currentResults = []
+    this.#currentDatabaseError = null
 
-    this.#protocolParser.parse(data, (msg) => {
+    const data = await this.execProtocolRaw(message, { syncToFs })
+
+    const databaseError = this.#currentDatabaseError
+    this.#currentThrowOnError = false
+    this.#currentOnNotice = undefined
+    this.#currentDatabaseError = null
+    const result = { messages: this.#currentResults, data }
+    this.#currentResults = []
+
+    if (throwOnError && databaseError) {
+      this.#protocolParser = new ProtocolParser() // Reset the parser
+      throw databaseError
+    }
+
+    return result
+  }
+
+  /**
+   * Execute a postgres wire protocol message
+   * @param message The postgres wire protocol message to execute
+   * @returns The parsed results of the query
+   */
+  async execProtocolStream(
+    message: Uint8Array,
+    { syncToFs, throwOnError = true, onNotice }: ExecProtocolOptions = {},
+  ): Promise<BackendMessage[]> {
+    this.#currentThrowOnError = throwOnError
+    this.#currentOnNotice = onNotice
+    this.#currentResults = []
+    this.#currentDatabaseError = null
+
+    this.#keepRawResponse = false
+
+    await this.execProtocolRaw(message, { syncToFs })
+
+    this.#keepRawResponse = true
+
+    const databaseError = this.#currentDatabaseError
+    this.#currentThrowOnError = false
+    this.#currentOnNotice = undefined
+    this.#currentDatabaseError = null
+    const result = this.#currentResults
+    this.#currentResults = []
+
+    if (throwOnError && databaseError) {
+      this.#protocolParser = new ProtocolParser() // Reset the parser
+      throw databaseError
+    }
+
+    return result
+  }
+
+  #parse(msg: BackendMessage) {
+    // keep the existing logic of throwing the first db exception
+    // as soon as there is a db error, we're not interested in the remaining data
+    // but since the parser is plugged into the pglite_write callback, we can't just throw
+    // and need to ack the messages received from the db
+    if (!this.#currentDatabaseError) {
       if (msg instanceof DatabaseError) {
-        this.#protocolParser = new ProtocolParser() // Reset the parser
-        if (throwOnError) {
-          throw msg
+        if (this.#currentThrowOnError) {
+          this.#currentDatabaseError = msg
         }
         // TODO: Do we want to wrap the error in a custom error?
       } else if (msg instanceof NoticeMessage) {
@@ -715,8 +792,8 @@ export class PGlite
           // Notice messages are warnings, we should log them
           console.warn(msg)
         }
-        if (onNotice) {
-          onNotice(msg)
+        if (this.#currentOnNotice) {
+          this.#currentOnNotice(msg)
         }
       } else if (msg instanceof CommandCompleteMessage) {
         // Keep track of the transaction state
@@ -743,10 +820,8 @@ export class PGlite
           queueMicrotask(() => cb(msg.channel, msg.payload))
         })
       }
-      results.push(msg)
-    })
-
-    return { messages: results, data }
+      this.#currentResults.push(msg)
+    }
   }
 
   /**