+++ packages/pglite/tsup.config.ts
@@ -26,6 +26,8 @@ const entryPoints = [
   'src/vector/index.ts',
   'src/pg_ivm/index.ts',
   'src/postgis/index.ts',
+  'src/pgtap/index.ts',
+  'src/pg_uuidv7/index.ts',
   'src/worker/index.ts',
 ]