+++ packages/pglite-sync/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @electric-sql/pglite-sync
 
+## 0.4.0
+
+### Minor Changes
+
+- 408500c: Allow passing in onError to 'syncShapesToTables'
+
+### Patch Changes
+
+- Updated dependencies [8785034]
+- Updated dependencies [90cfee8]
+  - @electric-sql/pglite@0.3.14
+
+## 0.3.17
+
+### Patch Changes
+
+- Updated dependencies [ad3d0d8]
+  - @electric-sql/pglite@0.3.13
+
+## 0.3.16
+
+### Patch Changes
+
+- Updated dependencies [ce0e74e]
+  - @electric-sql/pglite@0.3.12
+
+## 0.3.15
+
+### Patch Changes
+
+- Updated dependencies [9a104b9]
+  - @electric-sql/pglite@0.3.11
+
 ## 0.3.14
 
 ### Patch Changes